requirements.txt
exif
matplotlib
numpy
opencv-python
pandas
Pillow
xmltodict